+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changes:
 -   Alert component add
 -   AccountsManager widget styles fix
 -   Link component props fix
+-   New icons
 
 ## 0.0.35 August 19, 2022